+++ src/budgetingpanel.cpp
@@ -67,7 +67,7 @@ EVT_LIST_ITEM_SELECTED(wxID_ANY, budgetingListCtrl::OnListItemSelected)
 EVT_LIST_ITEM_ACTIVATED(wxID_ANY, budgetingListCtrl::OnListItemActivated)
 wxEND_EVENT_TABLE()
 /*******************************************************/
-mmBudgetingPanel::mmBudgetingPanel(int budgetYearID
+mmBudgetingPanel::mmBudgetingPanel(int64 budgetYearID
     , wxWindow *parent, mmGUIFrame *frame
     , wxWindowID winid
     , const wxPoint& pos, const wxSize& size
@@ -304,7 +304,7 @@ void mmBudgetingPanel::sortTable()
     //TODO: Sort budget panel
 }
 
-bool mmBudgetingPanel::DisplayEntryAllowed(int categoryID, int subcategoryID)
+bool mmBudgetingPanel::DisplayEntryAllowed(int64 categoryID, int64 subcategoryID)
 {
     bool result = false;
 
@@ -574,7 +574,7 @@ wxString mmBudgetingPanel::getItem(long item, long column)
             ? budget_[item].first : budget_[item].second);
         if (category) {
             wxString name = category->CATEGNAME;
-            for (int i = displayDetails_[category->CATEGID].first; i > 0; i--) {
+            for (int64 i = displayDetails_[category->CATEGID].first; i > 0; i--) {
                 name.Prepend("    ");
             }
             return name;
